Move AppProvider to providers/ directory
